+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Audio Player project will be documented in this file.
 ### Added
 - gapless playback through preloading next track
 - store and display track comments
+- playback speed toggle button
 
 ### Changed
 - refactored MusicController into service and mapper